+++ scripts/update-citation.R
@@ -3,5 +3,6 @@ cff <- cff_create(
   dependencies = FALSE
 )
 
+fs::file_delete("inst/CITATION")
 cff_write(cff)
 cff_write_citation(cff, file = "inst/CITATION")