+++ NEWS.md
@@ -1,5 +1,6 @@
 # fluxnet (development version)
 
+* Changed default cache age for `flux_listall()` to 1 day.
 * `flux_download()` now retries failed downloads once and `overwrite = FALSE` no longer skips downloading corrupted or partial downloads.
 * Added `flux_read()` for reading in FLUXNET data in a manifest.
 * Added `flux_discover_files()` for creating a file manifest.