+++ NEWS.md
@@ -1,6 +1,7 @@
 # fluxnet (development version)
 
-* Added `flux_map_sites()`
+* Added `flux_qc()`, a function to flag overly gapfilled rows of annual data.
+* Added `flux_map_sites()`.
 * `flux_download()` prints a warning when downloads fail.
 * Changed default cache age for `flux_listall()` to 1 day.
 * `flux_download()` now retries failed downloads once and `overwrite = FALSE` no longer skips downloading corrupted or partial downloads.