+++ NEWS.md
@@ -1,6 +1,6 @@
 # fluxnet (development version)
 
-* Added `flux_qc()`, a function to flag overly gapfilled rows of annual data.
+* Added `flux_qc()`, a function to flag overly gapfilled rows of aggregated (not hourly) data.
 * Added `flux_map_sites()`.
 * `flux_download()` now uses `httr2` rather than `curl` for downloading files.  This fixed a bug where downloads were failing when attempting to download large numbers of sites at once. 
 * Added `flux_map_sites()`